+++ common/src/main/java/com/cozary/ore_creeper/init/ModEntityTypes.java
@@ -48,6 +48,9 @@ public class ModEntityTypes {
     public static final RegistryObject<EntityType<RedstoneCreeperEntity>> REDSTONE_CREEPER = registerEntitiesList("redstone_creeper", () -> EntityType.Builder.of(RedstoneCreeperEntity::new, MobCategory.MONSTER)
             .sized(0.6F, 1.7F).clientTrackingRange(8)
             .build(new ResourceLocation(OreCreeper.MOD_ID, "redstone_creeper").toString()));
+    public static final RegistryObject<EntityType<AncientDebrisCreeperEntity>> ANCIENT_DEBRIS_CREEPER = registerEntitiesList("ancient_debris_creeper", () -> EntityType.Builder.of(AncientDebrisCreeperEntity::new, MobCategory.MONSTER)
+            .sized(0.6F, 1.7F).clientTrackingRange(8)
+            .build(ResourceLocation.fromNamespaceAndPath(OreCreeper.MOD_ID, "ancient_debris_creeper").toString()));
     public static final RegistryObject<EntityType<OrePrimedTnt>> ORE_PRIMED_TNT = ENTITY_TYPES.register("ore_primed_tnt", () -> EntityType.Builder.<OrePrimedTnt>of(OrePrimedTnt::new, MobCategory.MISC)
             .fireImmune().sized(0.98F, 0.98F).clientTrackingRange(10).updateInterval(10)
             .build(new ResourceLocation(OreCreeper.MOD_ID, "ore_primed_tnt").toString()));