+++ common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Changed
+
+- Spawn rate halved
+
 ### Fixed
 
 - recipes -> recipe folder rename